rename common_criteria modules to cc
