+++ src/sec_certs/dataset/__init__.py
@@ -1,6 +1,6 @@
 """This package exposes Datasets of various Samples, both primary (Common Criteria, FIPS) and auxillary (CVEs, CPEs, ...)"""
 
-from sec_certs.dataset.common_criteria import CCDataset, CCDatasetMaintenanceUpdates
+from sec_certs.dataset.cc import CCDataset, CCDatasetMaintenanceUpdates, CCSchemeDataset
 from sec_certs.dataset.cpe import CPEDataset
 from sec_certs.dataset.cve import CVEDataset
 from sec_certs.dataset.fips import FIPSDataset
@@ -12,6 +12,7 @@ from sec_certs.dataset.protection_profile import ProtectionProfileDataset
 __all__ = [
     "CCDataset",
     "CCDatasetMaintenanceUpdates",
+    "CCSchemeDataset",
     "CPEDataset",
     "CVEDataset",
     "FIPSDataset",
