+++ src/sec_certs/sample/__init__.py
@@ -2,9 +2,9 @@
 like CPE, CVE, etc. The objects mostly hold data and allow for serialization, but can also perform some basic transformations.
 """
 
+from sec_certs.sample.cc import CommonCriteriaCert
 from sec_certs.sample.cc_certificate_id import CertificateId
 from sec_certs.sample.cc_maintenance_update import CommonCriteriaMaintenanceUpdate
-from sec_certs.sample.common_criteria import CommonCriteriaCert
 from sec_certs.sample.cpe import CPE, cached_cpe
 from sec_certs.sample.cve import CVE
 from sec_certs.sample.fips import FIPSCertificate
