+++ tests/data/cc/analysis/transitive_vulnerability_dataset.json
@@ -1,5 +1,5 @@
 {
-  "_type": "sec_certs.dataset.common_criteria.CCDataset",
+  "_type": "sec_certs.dataset.cc.CCDataset",
   "state": {
     "_type": "sec_certs.dataset.dataset.Dataset.DatasetInternalState",
     "meta_sources_parsed": false,
@@ -14,7 +14,7 @@
   "n_certs": 3,
   "certs": [
     {
-      "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert",
+      "_type": "sec_certs.sample.cc.CommonCriteriaCert",
       "dgst": "d0705c9e6fbaeba3",
       "status": "active",
       "category": "Operating Systems",
@@ -53,7 +53,7 @@
         "elements": []
       },
       "state": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.InternalState",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.InternalState",
         "st_download_ok": true,
         "report_download_ok": true,
         "st_convert_ok": true,
@@ -66,7 +66,7 @@
         "report_txt_hash": "9d360141a98e764b15855f519b456c4e4639f993c4f8b5ab67e9c8ae7fbfc9e4"
       },
       "pdf_data": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.PdfData",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.PdfData",
         "report_metadata": {
           "pdf_file_size_bytes": 1235750,
           "pdf_is_encrypted": false,
@@ -1061,7 +1061,7 @@
         }
       },
       "heuristics": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.Heuristics",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.Heuristics",
         "extracted_versions": {
           "_type": "Set",
           "elements": [
@@ -1329,7 +1329,7 @@
       }
     },
     {
-      "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert",
+      "_type": "sec_certs.sample.cc.CommonCriteriaCert",
       "dgst": "011796336c7b94de",
       "status": "archived",
       "category": "Operating Systems",
@@ -1358,7 +1358,7 @@
         "elements": []
       },
       "state": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.InternalState",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.InternalState",
         "st_download_ok": true,
         "report_download_ok": true,
         "st_convert_ok": true,
@@ -1371,7 +1371,7 @@
         "report_txt_hash": "dd120ba7667c2385839c96ee70c56f2a4d464fc95e3ea2818d31b3347d06fd4f"
       },
       "pdf_data": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.PdfData",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.PdfData",
         "report_metadata": {
           "pdf_file_size_bytes": 1178202,
           "pdf_is_encrypted": false,
@@ -2014,7 +2014,7 @@
         }
       },
       "heuristics": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.Heuristics",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.Heuristics",
         "extracted_versions": {
           "_type": "Set",
           "elements": [
@@ -2269,7 +2269,7 @@
       }
     },
     {
-      "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert",
+      "_type": "sec_certs.sample.cc.CommonCriteriaCert",
       "dgst": "ebc77980250ee68f",
       "status": "active",
       "category": "Operating Systems",
@@ -2308,7 +2308,7 @@
         "elements": []
       },
       "state": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.InternalState",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.InternalState",
         "st_download_ok": true,
         "report_download_ok": true,
         "st_convert_ok": true,
@@ -2321,7 +2321,7 @@
         "report_txt_hash": "0a7c65e3d11f082c8f75aba7de0079c0b1aa5e67bb28d4635cbcaa4cd200d1c2"
       },
       "pdf_data": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.PdfData",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.PdfData",
         "report_metadata": {
           "pdf_file_size_bytes": 1932018,
           "pdf_is_encrypted": false,
@@ -3380,7 +3380,7 @@
         }
       },
       "heuristics": {
-        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.Heuristics",
+        "_type": "sec_certs.sample.cc.CommonCriteriaCert.Heuristics",
         "extracted_versions": {
           "_type": "Set",
           "elements": [
