+++ tests/data/cc/dataset/toy_dataset.json
@@ -1,5 +1,5 @@
 {
-    "_type": "sec_certs.dataset.common_criteria.CCDataset",
+    "_type": "sec_certs.dataset.cc.CCDataset",
     "state": {
         "_type": "sec_certs.dataset.dataset.Dataset.DatasetInternalState",
         "meta_sources_parsed": true,
@@ -15,7 +15,7 @@
     "n_certs": 3,
     "certs": [
         {
-            "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert",
+            "_type": "sec_certs.sample.cc.CommonCriteriaCert",
             "dgst": "309ac2fd7f2dcf17",
             "status": "active",
             "category": "Access Control Devices and Systems",
@@ -44,7 +44,7 @@
                 "elements": []
             },
             "state": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.InternalState",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.InternalState",
                 "st_download_ok": false,
                 "report_download_ok": false,
                 "st_convert_garbage": false,
@@ -59,7 +59,7 @@
                 "report_txt_hash": null
             },
             "pdf_data": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.PdfData",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.PdfData",
                 "report_metadata": null,
                 "st_metadata": null,
                 "report_frontpage": null,
@@ -70,7 +70,7 @@
                 "st_filename": null
             },
             "heuristics": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.Heuristics",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.Heuristics",
                 "extracted_versions": null,
                 "cpe_matches": null,
                 "verified_cpe_matches": null,
@@ -97,7 +97,7 @@
             }
         },
         {
-            "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert",
+            "_type": "sec_certs.sample.cc.CommonCriteriaCert",
             "dgst": "8cf86948f02f047d",
             "status": "active",
             "category": "Access Control Devices and Systems",
@@ -131,7 +131,7 @@
                 "elements": []
             },
             "state": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.InternalState",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.InternalState",
                 "st_download_ok": false,
                 "report_download_ok": false,
                 "st_convert_garbage": false,
@@ -146,7 +146,7 @@
                 "report_txt_hash": null
             },
             "pdf_data": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.PdfData",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.PdfData",
                 "report_metadata": null,
                 "st_metadata": null,
                 "report_frontpage": null,
@@ -157,7 +157,7 @@
                 "st_filename": null
             },
             "heuristics": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.Heuristics",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.Heuristics",
                 "extracted_versions": null,
                 "cpe_matches": null,
                 "verified_cpe_matches": null,
@@ -184,7 +184,7 @@
             }
         },
         {
-            "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert",
+            "_type": "sec_certs.sample.cc.CommonCriteriaCert",
             "dgst": "8a5e6bcda602920c",
             "status": "active",
             "category": "Boundary Protection Devices and Systems",
@@ -217,7 +217,7 @@
                 "_type": "Set",
                 "elements": [
                     {
-                        "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.MaintenanceReport",
+                        "_type": "sec_certs.sample.cc.CommonCriteriaCert.MaintenanceReport",
                         "maintenance_date": "2019-08-26",
                         "maintenance_title": "Fortinet FortiGate w/ FortiOS v5.6.7 Build 6022",
                         "maintenance_report_link": "https://www.commoncriteriaportal.org/files/epfiles/383-7-159%20MR%20v1.0e.pdf",
@@ -226,7 +226,7 @@
                 ]
             },
             "state": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.InternalState",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.InternalState",
                 "st_download_ok": false,
                 "report_download_ok": false,
                 "st_convert_garbage": false,
@@ -241,7 +241,7 @@
                 "report_txt_hash": null
             },
             "pdf_data": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.PdfData",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.PdfData",
                 "report_metadata": null,
                 "st_metadata": null,
                 "report_frontpage": null,
@@ -252,7 +252,7 @@
                 "st_filename": null
             },
             "heuristics": {
-                "_type": "sec_certs.sample.common_criteria.CommonCriteriaCert.Heuristics",
+                "_type": "sec_certs.sample.cc.CommonCriteriaCert.Heuristics",
                 "extracted_versions": null,
                 "cpe_matches": null,
                 "verified_cpe_matches": null,
